- Remove tag button from search menu (feature will be reintroduced later) - Code cleaning & commenting
